scope and annotations defined on layers
scripts updated for build dir changed

html.schtml dependency from js.core broke js.test.simpleBinding which requires that
we do not have the <script> tags loaded in the schtml page for index.schtml.  I think
a more robust way to fix this would be to have index.schmtl just extend a hard-coded HtmlPage
class so it does not inherit the one but there's a bug in that case now.  It requires
the extends class to extend the same as the previous class which is not a limitation we
have in SC.